# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Film-collection API",
    "contact": {
      "name": "Yuliia Kopysova",
      "email": "julykopysova97@b.com"
    }
  },
  "host": "172.20.0.3:5000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
         "/genres": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Get genres",
                "description": ":return: JSON",
                "tags": [
                    "Genres"
                ]
            },
            "post": {
                "summary": "API Endpoint for create Genre by admin",
                "description": "Create new Genre by admin",
                "tags": [
                            "Genres"
                        ],
                "parameters": [
                  {
                    "name": "genre",
                    "description": "Genre params for creating",
                    "in": "body",
                    "schema" : {

                        }
                  }
                ],
                "responses": {
                  "200": {
                    "description": "New genre has created"
                  },
                  "401": {
                    "description": "User is not admin"
                  }
                }
            }
        },
    "/genres/{genre_id}": {
            "parameters": [
                {
                    "name": "genre_id",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                }
            ],
            "delete": {
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "204": {
                        "description": "Genre was deleted"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "User is not admin"
                    }
                },
                "summary": "Delete genre",
                "description": ":param genre_id: id of genre\n:return: Response",
                "tags": [
                    "Genres"
                ]
            },
            "patch": {
              "parameters": [
                  {
                    "name": "genre",
                    "description": "Genre params for updating",
                    "in": "body",
                    "schema" : {

                        }
                  }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Not authorized"
                    }
                },
                "summary": "Updates a genre",
                "description": ":param genre_id: id of genre\n:return: JSON",
                "tags": [
                    "Genres"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Get one genre",
                "description": ":param genre_id: id of genre\n:return: JSON",
                "tags": [
                    "Genres"
                ]
            }
        },
...
    "/films": {
            "get": {
               "parameters": [
                  {
                    "name": "ordering",
                    "description": "By what film the ordering will be execute",
                    "in": "query",
                    "schema" : {

                        }
                  }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Get films",
                "description": ":return: JSON",
                "tags": [
                    "Films"
                ]
            },
            "post": {
                "summary": "API Endpoint for create film by user",
                "description": "Create new film",
                "tags": [
                            "Films"
                        ],
                "parameters": [
                  {
                    "name": "film",
                    "description": "Film  params for adding",
                    "in": "body",
                    "schema" : {

                        }
                  }
                ],
                "responses": {
                  "200": {
                    "description": "New film has added"
                  },
                  "401": {
                    "description": "User isn't authenticated, Incorrect data"
                  }
                }
            }
        },
    "/films/{film_id}": {
            "parameters": [
                {
                    "name": "film_id",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                }
            ],
            "delete": {
                "responses": {
                    "200": {
                        "description": "Film was deleted"
                    },
                    "401": {
                        "description": "You aren't admin or user who added this film\nFilm doesn't exist"
                    }
                },
                "summary": "Delete film",
                "description": ":param film_id: id of film\n:return: Response",
                "tags": [
                    "Films"
                ]
            },
            "patch": {
              "parameters": [
                  {
                    "name": "film",
                    "description": "Film params for updating",
                    "in": "body",
                    "schema" : {

                        }
                  }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "You aren't admin or user who added this film"
                    }
                },
                "summary": "Updates a film",
                "description": ":param film_id: id of film\n:return: JSON",
                "tags": [
                    "Films"
                ]
            },
            "get": {
                "responses": {
                  "200": {
                        "description": "Success"
                    }
                },
                "summary": "Get one film",
                "description": ":param film_id: id of film\n:return: JSON",
                "tags": [
                    "Films"
                ]
            }
        },
...
  }
}